+++ doc/changes.rst
@@ -2,6 +2,18 @@ Changelog
 =========
 If you are upgrading an existing installation, read :ref:`the instructions <upgrade>`
 
+1.3.8 - 2022/8/11
+-----------------
+* Add Redis Cluster support (:pr:`309`)
+
+1.3.7 - 2022/7/13
+-----------------
+* Fix login issues in the web interface (:issue:`307`)
+
+1.3.6 - 2022/7/7
+----------------
+* Allow `/` to serve packages the same as `/simple` (:issue:`305`)
+
 1.3.5 - 2022/6/20
 -----------------
 * Handle all errors when fetching from upstream (:pr:`299`)